+++ src/main/java/com/cargoguardian/model/Shipment.java
@@ -43,9 +43,9 @@ public class Shipment {
         this.liquidItemContentPercentage = newPercent;
     }
 
-    public double getSafetyScore()
+    public int getSafetyScore()
     {
-        return 0.0;
+        return 0;
     }
